+++ packages/svelte/src/lib/container/SvelteFlow/types.ts
@@ -18,8 +18,7 @@ import type {
   ColorMode,
   OnConnect,
   OnConnectStart,
-  OnConnectEnd,
-  OnBeforeDelete
+  OnConnectEnd
 } from '@xyflow/system';
 
 import type {
@@ -31,7 +30,8 @@ import type {
   DefaultEdgeOptions,
   FitViewOptions,
   OnDelete,
-  OnEdgeCreate
+  OnEdgeCreate,
+  OnBeforeDelete
 } from '$lib/types';
 import type { Writable } from 'svelte/store';
 
@@ -135,11 +135,11 @@ export type SvelteFlowProps = DOMAttributes<HTMLDivElement> & {
    */
   nodeDragThreshold?: number;
   /** Minimum zoom level
-   * @default 0.1
+   * @default 0.5
    */
   minZoom?: number;
   /** Maximum zoom level
-   * @default 1
+   * @default 2
    */
   maxZoom?: number;
   /** Sets the initial position and zoom of the viewport.