+++ packages/svelte/src/lib/hooks/useHandleConnections.ts
@@ -1,15 +1,16 @@
 import { derived } from 'svelte/store';
-import { areConnectionMapsEqual, type Connection, type HandleType } from '@xyflow/system';
+import { areConnectionMapsEqual, type HandleConnection, type HandleType } from '@xyflow/system';
 
 import { useStore } from '$lib/store';
+import { getContext } from 'svelte';
 
 export type useHandleConnectionsParams = {
-  nodeId: string;
   type: HandleType;
+  nodeId?: string;
   id?: string | null;
 };
 
-const initialConnections: Connection[] = [];
+const initialConnections: HandleConnection[] = [];
 
 /**
  *  Hook to check if a <Handle /> is connected to another <Handle /> and get the connections.
@@ -20,14 +21,18 @@ const initialConnections: Connection[] = [];
  * @param param.id - the handle id (this is only needed if the node has multiple handles of the same type)
  * @returns an array with connections
  */
-export function useHandleConnections({ nodeId, type, id = null }: useHandleConnectionsParams) {
+export function useHandleConnections({ type, nodeId, id = null }: useHandleConnectionsParams) {
   const { edges, connectionLookup } = useStore();
-  let prevConnections: Map<string, Connection> | undefined = undefined;
+
+  const _nodeId = getContext<string>('svelteflow__node_id');
+  const currentNodeId = nodeId ?? _nodeId;
+
+  let prevConnections: Map<string, HandleConnection> | undefined = undefined;
 
   return derived(
     [edges, connectionLookup],
     ([, connectionLookup], set) => {
-      const nextConnections = connectionLookup.get(`${nodeId}-${type}-${id || null}`);
+      const nextConnections = connectionLookup.get(`${currentNodeId}-${type}-${id || null}`);
 
       if (!areConnectionMapsEqual(nextConnections, prevConnections)) {
         prevConnections = nextConnections;