+++ packages/system/src/types/general.ts
@@ -28,6 +28,10 @@ export type Connection = {
   targetHandle: string | null;
 };
 
+export type HandleConnection = Connection & {
+  edgeId: string;
+};
+
 export type ConnectionStatus = 'valid' | 'invalid';
 
 export enum ConnectionMode {
@@ -123,11 +127,7 @@ export type SelectionRect = Rect & {
 
 export type OnError = (id: string, message: string) => void;
 
-export type UpdateNodePositions = (
-  dragItems: NodeDragItem[] | NodeBase[],
-  positionChanged?: boolean,
-  dragging?: boolean
-) => void;
+export type UpdateNodePositions = (dragItems: NodeDragItem[] | NodeBase[], dragging?: boolean) => void;
 export type PanBy = (delta: XYPosition) => boolean;
 
 export type UpdateConnection = (params: {
@@ -140,7 +140,7 @@ export type UpdateConnection = (params: {
 export type ColorModeClass = 'light' | 'dark';
 export type ColorMode = ColorModeClass | 'system';
 
-export type ConnectionLookup = Map<string, Map<string, Connection>>;
+export type ConnectionLookup = Map<string, Map<string, HandleConnection>>;
 
 export type OnBeforeDeleteBase<NodeType extends NodeBase = NodeBase, EdgeType extends EdgeBase = EdgeBase> = ({
   nodes,